feat: split layout from index.tsx
