+++ packages/www/pages/index.tsx
@@ -1,44 +1,9 @@
 /* eslint-disable max-len */
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react';
-import Link from 'next/link';
 import { centerAbsolute, Button, theme as t } from '@codement/ui';
 import styled from 'styled-components';
-import People from '../public/people.svg';
-import Logo from '../public/logo.svg';
-
-const StyledHomePage = styled.div`
-  background: linear-gradient(180deg, ${t.color('white')}, ${t.colors.primary[100]});
-  height: 100vh;
-`;
-
-const StyledNav = styled.nav`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: ${t.size('lg')} ${t.size('giant')};
-`;
-
-const StyledNavList = styled.ul`
-  li {
-    display: inline-block;
-
-    :not(:last-of-type) {
-      margin: 0 ${t.size('lg')} 0 0;
-    }
-  }
-`;
-
-const StyledAnchor = styled.a`
-  color: ${t.colors.grey[700]};
-  font-weight: ${t.fontWeight.heavy};
-  text-transform: uppercase;
-  cursor: pointer;
-
-  &:hover {
-    color: ${t.color('primary')};
-  }
-`;
+import Layout from '../components/layout';
 
 const StyledHeroText = styled.div`
   width: 600px;
@@ -58,35 +23,10 @@ const StyledHeroText = styled.div`
   }
 `;
 
-const StyledPeople = styled(People)`
-  position: fixed;
-  left: -${t.size('tiny')};
-  bottom: -${t.size('tiny')};
-`;
 
 export default function Home() {
   return (
-    <StyledHomePage>
-      <StyledNav>
-        <Logo />
-        <StyledNavList>
-          <li>
-            <Link href="/pricing">
-              <StyledAnchor>Pricing</StyledAnchor>
-            </Link>
-          </li>
-          <li>
-            <Link href="/signup">
-              <StyledAnchor>Signup</StyledAnchor>
-            </Link>
-          </li>
-          <li>
-            <Link href="/login">
-              <StyledAnchor>Login</StyledAnchor>
-            </Link>
-          </li>
-        </StyledNavList>
-      </StyledNav>
+    <Layout people>
       <StyledHeroText>
         <h2>
           Coding got you lost? We&apos;re here to help!
@@ -98,7 +38,6 @@ export default function Home() {
           Learn coding for free
         </Button>
       </StyledHeroText>
-      <StyledPeople />
-    </StyledHomePage>
+    </Layout>
   );
 }
